Adds keyboard management
