+++ piwigo/Settings/Extensions/SettingsViewController+Albums.swift
@@ -50,6 +50,53 @@ extension SettingsViewController: DefaultAlbumThumbnailSizeDelegate {
 }
 
 
+// MARK: - SelectCategoryDelegate Methods
+extension SettingsViewController: SelectCategoryDelegate {
+    func didSelectCategory(withId categoryId: Int32) {
+        // Do nothing if new default album is unknown or unchanged
+        guard categoryId != Int32.min,
+              categoryId != AlbumVars.shared.defaultCategory else {
+            return
+        }
+
+        // Save new choice
+        AlbumVars.shared.defaultCategory = categoryId
+
+        // Change album name in row
+        let indexPath = IndexPath(row: 0, section: SettingsSection.albums.rawValue)
+        if let indexPaths = settingsTableView.indexPathsForVisibleRows, indexPaths.contains(indexPath),
+           let cell = settingsTableView.cellForRow(at: indexPath) as? LabelTableViewCell {
+            cell.detailLabel.text = defaultAlbumName()
+        }
+
+        // Switch to new default album
+        settingsDelegate?.didChangeDefaultAlbum()
+    }
+    
+    func defaultAlbumName() -> String {
+        var rootName: String
+        if view.bounds.size.width > 375 {
+            rootName = NSLocalizedString("categorySelection_root", comment: "Root Album")
+        } else {
+            rootName = NSLocalizedString("categorySelection_root<375pt", comment: "Root")
+        }
+
+        // Root album?
+        if AlbumVars.shared.defaultCategory == 0 {
+            return rootName
+        }
+        
+        // Default album…
+        if let album = albumProvider.getAlbum(ofUser: user, withId: AlbumVars.shared.defaultCategory),
+           album.name.isEmpty == false {
+            return album.name
+        } else {
+            return NSLocalizedString("categorySelection_title", comment: "Album")
+        }
+    }
+}
+
+
 // MARK: - CategorySortDelegate Methods
 extension SettingsViewController: CategorySortDelegate {
     func didSelectCategorySortType(_ sortType: pwgImageSort) {
