+++ piwigo/Settings/Extensions/SettingsViewController+UITableView.swift
@@ -1,0 +1,421 @@
+//
+//  SettingsViewController+UITableView.swift
+//  piwigo
+//
+//  Created by Eddy Lelièvre-Berna on 01/01/2024.
+//  Copyright © 2024 Piwigo.org. All rights reserved.
+//
+
+import MessageUI
+import piwigoKit
+
+extension SettingsViewController: UITableViewDelegate
+{
+    // MARK: - Headers
+    private func getContentOfHeader(inSection section: Int) -> (String, String) {
+        // Header strings
+        var title = "", text = ""
+        switch activeSection(section) {
+        case .server:
+            title = String(format: "%@ %@",
+                           NSLocalizedString("settingsHeader_server", comment: "Piwigo Server"),
+                           NetworkVars.pwgVersion)
+            if (NetworkVars.serverProtocol == "http://") {
+                title += "\n"
+                text = NSLocalizedString("settingsHeader_notSecure", comment: "Website Not Secure!")
+            }
+            if NetworkVars.pwgVersion.compare(NetworkVars.pwgRecentVersion, options: .numeric) == .orderedAscending {
+                if !title.contains("\n") { title += "\n" }
+                if !text.isEmpty { text += " — " }
+                text += NSLocalizedString("serverVersionOld_title", comment: "Server Update Available")
+            }
+        case .albums:
+            title = NSLocalizedString("tabBar_albums", comment: "Albums")
+        case .images:
+            title = NSLocalizedString("settingsHeader_images", comment: "Images")
+        case .imageUpload:
+            title = NSLocalizedString("settingsHeader_upload", comment: "Default Upload Settings")
+        case .appearance:
+            title = NSLocalizedString("settingsHeader_appearance", comment: "Appearance")
+        case .privacy:
+            title = NSLocalizedString("settingsHeader_privacy", comment: "Privacy")
+        case .cache:
+            title = NSLocalizedString("settingsHeader_cache", comment: "Cache Settings")
+        case .about:
+            title = NSLocalizedString("settingsHeader_about", comment: "Information")
+        case .logout, .clear:
+            fallthrough
+        default:
+            break
+        }
+        return (title, text)
+    }
+    
+    func tableView(_ tableView: UITableView, heightForHeaderInSection section: Int) -> CGFloat {
+        let (title, text) = getContentOfHeader(inSection: section)
+        if title.isEmpty, text.isEmpty {
+            return CGFloat(1)
+        } else {
+            return TableViewUtilities.shared.heightOfHeader(withTitle: title, text: text,
+                                                            width: tableView.frame.size.width)
+        }
+    }
+    
+    func tableView(_ tableView: UITableView, viewForHeaderInSection section: Int) -> UIView? {
+        let (title, text) = getContentOfHeader(inSection: section)
+        return TableViewUtilities.shared.viewOfHeader(withTitle: title, text: text)
+    }
+    
+    
+    // MARK: - Rows
+    func tableView(_ tableView: UITableView, heightForRowAt indexPath: IndexPath) -> CGFloat {
+        return 44.0
+    }
+    
+    func tableView(_ tableView: UITableView, shouldHighlightRowAt indexPath: IndexPath) -> Bool {
+        var result = true
+        switch activeSection(indexPath.section) {
+            // MARK: Server
+        case .server /* Piwigo Server */:
+            result = false
+        case .logout /* Logout Button */:
+            result = true
+            
+            // MARK: Albums
+        case .albums /* Albums */:
+            switch indexPath.row {
+            case 0 /* Default album */, 1 /* Default Thumbnail File */:
+                result = true
+            default:
+                result = false
+            }
+            
+            // MARK: Images
+        case .images /* Images */:
+            switch indexPath.row {
+            case 0 /* Default Sort */,
+                1 /* Default Thumbnail File */,
+                4 /* Default Size of Previewed Images */:
+                result = true
+            default:
+                result = false
+            }
+            
+            // MARK: Upload Settings
+        case .imageUpload /* Default Upload Settings */:
+            var row = indexPath.row
+            row += (!user.hasAdminRights && (row > 0)) ? 1 : 0
+            row += (!UploadVars.resizeImageOnUpload && (row > 3)) ? 2 : 0
+            row += (!UploadVars.compressImageOnUpload && (row > 6)) ? 1 : 0
+            row += (!UploadVars.prefixFileNameBeforeUpload && (row > 8)) ? 1 : 0
+            row += (!NetworkVars.usesUploadAsync && (row > 10)) ? 1 : 0
+            switch row {
+            case 1  /* Privacy Level */,
+                4  /* Upload Photo Size */,
+                5  /* Upload Video Size */,
+                11 /* Auto upload */:
+                result = true
+            default:
+                result = false
+            }
+            
+            // MARK: Privacy
+        case .privacy   /* Privacy */:
+            result = true
+            
+            // MARK: Appearance
+        case .appearance /* Appearance */:
+            result = true
+            
+            // MARK: Cache
+        case .cache /* Cache Settings */:
+            result = false
+        case .clear /* Cache Settings */:
+            result = true
+            
+            // MARK: Information
+        case .about /* Information */:
+            switch indexPath.row {
+            case 1 /* Contact Us */:
+                result = MFMailComposeViewController.canSendMail() ? true : false
+            case 0 /* Twitter */,
+                2 /* Support Forum */,
+                3 /* Rate Piwigo Mobile */,
+                4 /* Translate Piwigo Mobile */,
+                5 /* Release Notes */,
+                6 /* Acknowledgements */,
+                7 /* Privacy Policy */:
+                result = true
+            default:
+                result = false
+            }
+        default:
+            result = false
+        }
+        return result
+    }
+    
+    
+    // MARK: - Footers
+    private func getContentOfFooter(inSection section: Int) -> String {
+        var footer = ""
+        switch activeSection(section) {
+        case .logout:
+            if NetworkVars.serverFileTypes.isEmpty == false {
+                footer = "\(NSLocalizedString("settingsFooter_formats", comment: "The server accepts the following file formats")): \(NetworkVars.serverFileTypes.replacingOccurrences(of: ",", with: ", "))."
+            }
+        case .about:
+            footer = NetworkVars.pwgStatistics
+        default:
+            footer = ""
+        }
+        return footer
+    }
+    
+    func tableView(_ tableView: UITableView, heightForFooterInSection section: Int) -> CGFloat {
+        let text = getContentOfFooter(inSection: section)
+        return TableViewUtilities.shared.heightOfFooter(withText: text,
+                                                        width: tableView.frame.width)
+    }
+    
+    func tableView(_ tableView: UITableView, viewForFooterInSection section: Int) -> UIView? {
+        let text = getContentOfFooter(inSection: section)
+        return TableViewUtilities.shared.viewOfFooter(withText: text, alignment: .center)
+    }
+    
+    
+    // MARK: - Cell Management
+    func tableView(_ tableView: UITableView, didSelectRowAt indexPath: IndexPath) {
+        tableView.deselectRow(at: indexPath, animated: true)
+        switch activeSection(indexPath.section) {
+        // MARK: Server
+        case .server /* Piwigo Server */:
+            break
+
+        // MARK: Logout
+        case .logout /* Logout */:
+            loginLogout()
+
+        // MARK: Albums
+        case .albums /* Albums */:
+            switch indexPath.row {
+            case 0 /* Default album */:
+                let categorySB = UIStoryboard(name: "SelectCategoryViewController", bundle: nil)
+                guard let categoryVC = categorySB.instantiateViewController(withIdentifier: "SelectCategoryViewController") as? SelectCategoryViewController else { return }
+                categoryVC.user = user
+                if categoryVC.setInput(parameter: AlbumVars.shared.defaultCategory,
+                                       for: .setDefaultAlbum) {
+                    categoryVC.delegate = self
+                    navigationController?.pushViewController(categoryVC, animated: true)
+                }
+            case 1 /* Thumbnail file selection */:
+                let defaultThumbnailSizeSB = UIStoryboard(name: "DefaultAlbumThumbnailSizeViewController", bundle: nil)
+                guard let defaultThumbnailSizeVC = defaultThumbnailSizeSB.instantiateViewController(withIdentifier: "DefaultAlbumThumbnailSizeViewController") as? DefaultAlbumThumbnailSizeViewController else { return }
+                defaultThumbnailSizeVC.delegate = self
+                navigationController?.pushViewController(defaultThumbnailSizeVC, animated: true)
+            default:
+                break
+            }
+
+        // MARK: Images
+        case .images /* Images */:
+            switch indexPath.row {
+            case 0 /* Sort method selection */:
+                let categorySB = UIStoryboard(name: "CategorySortViewController", bundle: nil)
+                guard let categoryVC = categorySB.instantiateViewController(withIdentifier: "CategorySortViewController") as? CategorySortViewController else {return }
+                categoryVC.sortDelegate = self
+                navigationController?.pushViewController(categoryVC, animated: true)
+            case 1 /* Thumbnail file selection */:
+                let defaultThumbnailSizeSB = UIStoryboard(name: "DefaultImageThumbnailSizeViewController", bundle: nil)
+                guard let defaultThumbnailSizeVC = defaultThumbnailSizeSB.instantiateViewController(withIdentifier: "DefaultImageThumbnailSizeViewController") as? DefaultImageThumbnailSizeViewController else { return }
+                defaultThumbnailSizeVC.delegate = self
+                navigationController?.pushViewController(defaultThumbnailSizeVC, animated: true)
+            case 4 /* Image file selection */:
+                let defaultImageSizeSB = UIStoryboard(name: "DefaultImageSizeViewController", bundle: nil)
+                guard let defaultImageSizeVC = defaultImageSizeSB.instantiateViewController(withIdentifier: "DefaultImageSizeViewController") as? DefaultImageSizeViewController else { return }
+                defaultImageSizeVC.delegate = self
+                navigationController?.pushViewController(defaultImageSizeVC, animated: true)
+            default:
+                break
+            }
+
+        // MARK: Upload Settings
+        case .imageUpload /* Default upload Settings */:
+            var row = indexPath.row
+            row += (!user.hasAdminRights && (row > 0)) ? 1 : 0
+            row += (!UploadVars.resizeImageOnUpload && (row > 3)) ? 2 : 0
+            row += (!UploadVars.compressImageOnUpload && (row > 6)) ? 1 : 0
+            row += (!UploadVars.prefixFileNameBeforeUpload && (row > 8)) ? 1 : 0
+            row += (!NetworkVars.usesUploadAsync && (row > 10)) ? 1 : 0
+            switch row {
+            case 1 /* Default privacy selection */:
+                let privacySB = UIStoryboard(name: "SelectPrivacyViewController", bundle: nil)
+                guard let privacyVC = privacySB.instantiateViewController(withIdentifier: "SelectPrivacyViewController") as? SelectPrivacyViewController else { return }
+                privacyVC.delegate = self
+                privacyVC.privacy = pwgPrivacy(rawValue: UploadVars.defaultPrivacyLevel) ?? .everybody
+                navigationController?.pushViewController(privacyVC, animated: true)
+            case 4 /* Upload Photo Size */:
+                let uploadPhotoSizeSB = UIStoryboard(name: "UploadPhotoSizeViewController", bundle: nil)
+                guard let uploadPhotoSizeVC = uploadPhotoSizeSB.instantiateViewController(withIdentifier: "UploadPhotoSizeViewController") as? UploadPhotoSizeViewController else { return }
+                uploadPhotoSizeVC.delegate = self
+                uploadPhotoSizeVC.photoMaxSize = UploadVars.photoMaxSize
+                navigationController?.pushViewController(uploadPhotoSizeVC, animated: true)
+            case 5 /* Upload Video Size */:
+                let uploadVideoSizeSB = UIStoryboard(name: "UploadVideoSizeViewController", bundle: nil)
+                guard let uploadVideoSizeVC = uploadVideoSizeSB.instantiateViewController(withIdentifier: "UploadVideoSizeViewController") as? UploadVideoSizeViewController else { return }
+                uploadVideoSizeVC.delegate = self
+                uploadVideoSizeVC.videoMaxSize = UploadVars.videoMaxSize
+                navigationController?.pushViewController(uploadVideoSizeVC, animated: true)
+            case 11 /* Auto Upload */:
+                let autoUploadSB = UIStoryboard(name: "AutoUploadViewController", bundle: nil)
+                guard let autoUploadVC = autoUploadSB.instantiateViewController(withIdentifier: "AutoUploadViewController") as? AutoUploadViewController else { return }
+                autoUploadVC.user = user
+                navigationController?.pushViewController(autoUploadVC, animated: true)
+            default:
+                break
+            }
+
+        // MARK: Privacy
+        case .privacy   /* Privacy */:
+            switch indexPath.row {
+            case 0 /* Clear cache */:
+                // Display numpad for setting up a passcode
+                let appLockSB = UIStoryboard(name: "LockOptionsViewController", bundle: nil)
+                guard let appLockVC = appLockSB.instantiateViewController(withIdentifier: "LockOptionsViewController") as? LockOptionsViewController else { return }
+                appLockVC.delegate = self
+                navigationController?.pushViewController(appLockVC, animated: true)
+            case 1 /* Clear Clipboard */:
+                // Display list of delays
+                let delaySB = UIStoryboard(name: "ClearClipboardViewController", bundle: nil)
+                guard let delayVC = delaySB.instantiateViewController(withIdentifier: "ClearClipboardViewController") as? ClearClipboardViewController else { return }
+                delayVC.delegate = self
+                navigationController?.pushViewController(delayVC, animated: true)
+            case 2 /* Share image metadata options */:
+                let metadataOptionsSB = UIStoryboard(name: "ShareMetadataViewController", bundle: nil)
+                guard let metadataOptionsVC = metadataOptionsSB.instantiateViewController(withIdentifier: "ShareMetadataViewController") as? ShareMetadataViewController else { return }
+                navigationController?.pushViewController(metadataOptionsVC, animated: true)
+
+            default:
+                break
+            }
+
+        // MARK: Appearance
+        case .appearance /* Appearance */:
+            if #available(iOS 13.0, *) {
+                let colorPaletteSB = UIStoryboard(name: "ColorPaletteViewController", bundle: nil)
+                guard let colorPaletteVC = colorPaletteSB.instantiateViewController(withIdentifier: "ColorPaletteViewController") as? ColorPaletteViewController else { return }
+                navigationController?.pushViewController(colorPaletteVC, animated: true)
+            } else {
+                let colorPaletteSB = UIStoryboard(name: "ColorPaletteViewControllerOld", bundle: nil)
+                guard let colorPaletteVC = colorPaletteSB.instantiateViewController(withIdentifier: "ColorPaletteViewControllerOld") as? ColorPaletteViewControllerOld else { return }
+                navigationController?.pushViewController(colorPaletteVC, animated: true)
+            }
+
+        // MARK: Cache
+        case .clear /* Cache Clear */:
+            switch indexPath.row {
+            case 0 /* Clear cache */:
+                // Determine position of cell in table view
+                let rowAtIndexPath = IndexPath(row: 0, section: SettingsSection.clear.rawValue)
+                let rectOfCellInTableView = settingsTableView?.rectForRow(at: rowAtIndexPath)
+
+                // Present list of actions
+                let alert = getClearCacheAlert()
+                alert.view.tintColor = .piwigoColorOrange()
+                if #available(iOS 13.0, *) {
+                    alert.overrideUserInterfaceStyle = AppVars.shared.isDarkPaletteActive ? .dark : .light
+                } else {
+                    // Fallback on earlier versions
+                }
+                alert.popoverPresentationController?.sourceView = settingsTableView
+                alert.popoverPresentationController?.permittedArrowDirections = [.up, .down]
+                alert.popoverPresentationController?.sourceRect = rectOfCellInTableView ?? CGRect.zero
+                present(alert, animated: true, completion: {
+                    // Bugfix: iOS9 - Tint not fully Applied without Reapplying
+                    alert.view.tintColor = .piwigoColorOrange()
+                })
+            default:
+                break
+            }
+
+        // MARK: Information
+        case .about /* About — Informations */:
+            switch indexPath.row {
+            case 0 /* Open @piwigo on Twitter */:
+                if let url = URL(string: NSLocalizedString("settings_twitterURL", comment: "https://twitter.com/piwigo")) {
+                    UIApplication.shared.open(url)
+                }
+            case 1 /* Prepare draft email */:
+                if MFMailComposeViewController.canSendMail() {
+                    let composeVC = MFMailComposeViewController()
+                    composeVC.mailComposeDelegate = self
+
+                    // Configure the fields of the interface.
+                    composeVC.setToRecipients([
+                    NSLocalizedString("contact_email", tableName: "PrivacyPolicy", bundle: Bundle.main, value: "", comment: "Contact email")
+                    ])
+
+                    // Collect version and build numbers
+                    let appVersionString = Bundle.main.object(forInfoDictionaryKey: "CFBundleShortVersionString") as? String
+                    let appBuildString = Bundle.main.object(forInfoDictionaryKey: "CFBundleVersion") as? String
+
+                    // Compile ticket number from current date
+                    let dateFormatter = DateFormatter()
+                    dateFormatter.dateFormat = "yyyyMMddHHmm"
+                    dateFormatter.locale = NSLocale(localeIdentifier: NetworkVars.language) as Locale
+                    let date = Date()
+                    let ticketDate = dateFormatter.string(from: date)
+
+                    // Set subject
+                    composeVC.setSubject("[Ticket#\(ticketDate)]: \(NSLocalizedString("settings_appName", comment: "Piwigo Mobile")) \(NSLocalizedString("settings_feedback", comment: "Feedback"))")
+
+                    // Collect system and device data
+                    let deviceModel = UIDevice.current.modelName
+                    let deviceOS = UIDevice.current.systemName
+                    let deviceOSversion = UIDevice.current.systemVersion
+
+                    // Set message body
+                    composeVC.setMessageBody("\(NSLocalizedString("settings_appName", comment: "Piwigo Mobile")) \(appVersionString ?? "") (\(appBuildString ?? ""))\n\(deviceModel ) — \(deviceOS) \(deviceOSversion)\n==============>>\n\n", isHTML: false)
+
+                    // Present the view controller modally.
+                    present(composeVC, animated: true)
+                }
+            case 2 /* Open Piwigo support forum webpage with default browser */:
+                if let url = URL(string: NSLocalizedString("settings_pwgForumURL", comment: "http://piwigo.org/forum")) {
+                    UIApplication.shared.open(url)
+                }
+            case 3 /* Open Piwigo App Store page for rating */:
+                // See https://itunes.apple.com/us/app/piwigo/id472225196?ls=1&mt=8
+                if let url = URL(string: "itms-apps://itunes.apple.com/app/piwigo/id472225196?action=write-review") {
+                    UIApplication.shared.open(url)
+                }
+            case 4 /* Open Piwigo Crowdin page for translating */:
+                if let url = URL(string: "https://crowdin.com/project/piwigo-mobile") {
+                    UIApplication.shared.open(url)
+                }
+            case 5 /* Open Release Notes page */:
+                let releaseNotesSB = UIStoryboard(name: "ReleaseNotesViewController", bundle: nil)
+                let releaseNotesVC = releaseNotesSB.instantiateViewController(withIdentifier: "ReleaseNotesViewController") as? ReleaseNotesViewController
+                if let releaseNotesVC = releaseNotesVC {
+                    navigationController?.pushViewController(releaseNotesVC, animated: true)
+                }
+            case 6 /* Open Acknowledgements page */:
+                let aboutSB = UIStoryboard(name: "AboutViewController", bundle: nil)
+                let aboutVC = aboutSB.instantiateViewController(withIdentifier: "AboutViewController") as? AboutViewController
+                if let aboutVC = aboutVC {
+                    navigationController?.pushViewController(aboutVC, animated: true)
+                }
+            case 7 /* Open Privacy Policy page */:
+                let privacyPolicySB = UIStoryboard(name: "PrivacyPolicyViewController", bundle: nil)
+                let privacyPolicyVC = privacyPolicySB.instantiateViewController(withIdentifier: "PrivacyPolicyViewController") as? PrivacyPolicyViewController
+                if let privacyPolicyVC = privacyPolicyVC {
+                    navigationController?.pushViewController(privacyPolicyVC, animated: true)
+                }
+            default:
+                break
+            }
+        default:
+            break
+        }
+    }
+}
